feat: entites 파일 분리 (Layer - Slices - Segments) 1. entites/user/model - user model 분리
