+++ src/entities/user/model/types.ts
@@ -1,0 +1,19 @@
+export interface User {
+  id: number;
+  username: string;
+  firstName: string;
+  lastName: string;
+  age: number;
+  email: string;
+  phone: string;
+  image: string;
+  address: {
+    address: string;
+    city: string;
+    state: string;
+  };
+  company: {
+    name: string;
+    title: string;
+  };
+}
