+++ src/pages/PostsManagerPage.tsx
@@ -12,6 +12,7 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import React, { useEffect } from 'react';
 
 import { Comment } from '../entities/comment';
+import { User } from '../entities/user';
 import {
   Button,
   Card,
@@ -40,7 +41,7 @@ import { highlightText } from '../shared/lib';
 import { useCommentsStore } from '../stores';
 import { usePostsStore } from '../stores/posts';
 import { useSearchStore } from '../stores/search';
-import { Post, UpdatePost, User } from '../stores/types';
+import { Post, UpdatePost } from '../stores/types';
 import { useUIStore } from '../stores/ui';
 import { useUsersStore } from '../stores/users';
